LPS-89423 Refactor
