+++ modules/apps/segments/segments-api/bnd.bnd
@@ -7,6 +7,7 @@ Export-Package:\
 	com.liferay.segments.criteria,\
 	com.liferay.segments.criteria.contributor,\
 	com.liferay.segments.exception,\
+	com.liferay.segments.field,\
 	com.liferay.segments.model,\
 	com.liferay.segments.odata.matcher,\
 	com.liferay.segments.odata.retriever,\
