+++ modules/apps/segments/segments-api/src/main/java/com/liferay/segments/field/Field.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.segments.criteria;
+package com.liferay.segments.field;
 
 import java.io.Serializable;
