+++ modules/apps/segments/segments-api/src/main/resources/com/liferay/segments/field/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
