+++ modules/apps/segments/segments-service/src/main/java/com/liferay/segments/internal/criteria/contributor/UserSegmentsCriteriaContributor.java
@@ -23,8 +23,8 @@ import com.liferay.portal.kernel.model.UserGroup;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.odata.entity.EntityModel;
 import com.liferay.segments.criteria.Criteria;
-import com.liferay.segments.criteria.Field;
 import com.liferay.segments.criteria.contributor.SegmentsCriteriaContributor;
+import com.liferay.segments.field.Field;
 import com.liferay.segments.internal.odata.entity.EntityModelFieldMapper;
 import com.liferay.segments.internal.odata.entity.UserEntityModel;
